test(dq-dashboard): apply review feedback on CertificationFilter spec
