+++ openmetadata-ui/src/main/resources/ui/playwright/utils/dataQuality.ts
@@ -386,3 +386,28 @@ export const verifyBundleSuitePageLoaded = async (
     )
     .toBe(expectedTestCaseCount);
 };
+
+/** A `dataQualityReport` call captured for assertion in tests. */
+export type CapturedReport = { url: string; q: string; index: string };
+
+/**
+ * Subscribes to every `/dataQualityReport` request fired by the page and
+ * returns a live array of (url, q, index). Useful for asserting which
+ * indices were queried and what filter the dashboard sent.
+ */
+export function captureReports(page: Page): CapturedReport[] {
+  const captured: CapturedReport[] = [];
+  page.on('request', (req) => {
+    const url = req.url();
+    if (!url.includes('/dataQualityReport')) {
+      return;
+    }
+    const u = new URL(url);
+    captured.push({
+      url,
+      q: u.searchParams.get('q') ?? '',
+      index: u.searchParams.get('index') ?? '',
+    });
+  });
+  return captured;
+}
